refactor(cycle/form): combine create and update form into a single component
