+++ src/components/Cycle/ActionModals/CreateCycleModal/messages.ts
@@ -8,6 +8,8 @@ type ActionMenuMessage =
   | 'successToastMessage'
   | 'existingUserErrorToastMessage'
   | 'unknownErrorToastMessage'
+  | 'closeModalButton'
+  | 'createCycleButton'
 
 export default defineMessages<ActionMenuMessage>({
   crateCycleModalTitle: {
@@ -50,5 +52,17 @@ export default defineMessages<ActionMenuMessage>({
     id: 'cGFQsk',
     description:
       'This message appears as an error toast when we have an unknown error while creating a new cycle',
+  },
+
+  closeModalButton: {
+    defaultMessage: 'Cancelar',
+    id: 'ej/Q3X',
+    description: 'The button to close the modal cycle',
+  },
+
+  createCycleButton: {
+    defaultMessage: 'Criar',
+    id: 'zWpOXa',
+    description: 'The button to create or edit a cycle on modal',
   },
 })
